+++ app/models/autotest_setting.rb
@@ -1,8 +1,14 @@
 class AutotestSetting < ApplicationRecord
   include AutomatedTestsHelper::AutotestApi
 
+  has_many :courses
+
   validates :url, presence: true
   before_create :register_autotester
+
+  # Column name used to identify this record if the primary identifier (id) cannot be relied on.
+  # For example, when unarchiving courses.
+  IDENTIFIER = 'url'.freeze
 
   private