+++ db/seeds.rb
@@ -1,9 +1,9 @@
 exit unless Rails.env.development?
 
 # clean up existing files first
-FileUtils.rm_rf(Repository::ROOT_DIR)
-FileUtils.rm_rf(Assignment::STARTER_FILES_DIR)
-FileUtils.rm_rf(TestRun::SETTINGS_FILES_DIR)
+FileUtils.rm_rf(Repository.root_dir)
+FileUtils.rm_rf(Assignment.starter_files_dir)
+FileUtils.rm_rf(TestRun.settings_files_dir)
 
 FileUtils.mkdir_p('tmp')