+++ spec/lib/repo/git_revision_spec.rb
@@ -1,7 +1,7 @@
 describe GitRevision do
   context 'with a git repo' do
     let(:repo) { build(:git_repository) }
-    before(:each) { FileUtils.rm_r(Dir.glob(File.join(Repository::ROOT_DIR, '*'))) }
+    before(:each) { FileUtils.rm_r(Dir.glob(File.join(Repository.root_dir, '*'))) }
     describe '#files_at_path' do
       # Commit a file named test in the workdir
       before(:each) do